Separate processing functionnalities in other library
