+++ calcli/application/main.cpp
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include <calcli/app/application.hpp>
+#include <calcli/application/application.hpp>
 
 
 int main()
