+++ calcli/processing/lexer.cpp
@@ -19,7 +19,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include <calcli/core/lexer.hpp>
+#include <calcli/processing/lexer.hpp>
 
 #include <cctype>
